requirements.txt
Flask==3.0.2
python-dotenv==1.0.0
console-menu